pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[project]
name = "freqsplit"
version = "0.1.0"
dependencies = [
    "aiohappyeyeballs==2.4.6",
    "aiohttp==3.11.13",
    "aiosignal==1.3.2",
    "amqp==5.3.1",
    "antlr4-python3-runtime==4.9.3",
    "appdirs==1.4.4",
    "asgiref==3.8.1",
    "asteroid==0.7.0",
    "asteroid-filterbanks==0.4.0",
    "attrs==25.1.0",
    "audioread==3.0.1",
    "billiard==4.2.1",
    "cached-property==2.0.1",
    "celery==5.4.0",
    "certifi==2025.1.31",
    "cffi==1.17.1",
    "charset-normalizer==3.4.1",
    "click==8.1.8",
    "click-didyoumean==0.3.1",
    "click-plugins==1.1.1",
    "click-repl==0.3.0",
    "cloudpickle==3.1.1",
    "contourpy==1.3.1",
    "cycler==0.12.1",
    "decorator==5.2.1",
    "DeepFilterLib==0.5.6",
    "DeepFilterNet==0.5.6",
    "demucs==4.0.1",
    "Django==5.1.6",
    "djangorestframework==3.15.2",
    "dora_search==0.1.12",
    "einops==0.8.1",
    "filelock==3.17.0",
    "fonttools==4.56.0",
    "frozenlist==1.5.0",
    "fsspec==2025.2.0",
    "huggingface-hub==0.29.1",
    "idna==3.10",
    "iniconfig==2.0.0",
    "Jinja2==3.1.5",
    "joblib==1.4.2",
    "julius==0.2.7",
    "kiwisolver==1.4.8",
    "kombu==5.4.2",
    "lameenc==1.8.1",
    "lazy_loader==0.4",
    "librosa==0.10.2.post1",
    "lightning-utilities==0.12.0",
    "llvmlite==0.44.0",
    "loguru==0.7.3",
    "MarkupSafe==3.0.2",
    "matplotlib==3.10.0",
    "mir_eval==0.8.2",
    "mpmath==1.3.0",
    "msgpack==1.1.0",
    "multidict==6.1.0",
    "networkx==3.4.2",
    "numba==0.61.0",
    "numpy==1.26.4",
    "nvidia-cublas-cu12==12.4.5.8",
    "nvidia-cuda-cupti-cu12==12.4.127",
    "nvidia-cuda-nvrtc-cu12==12.4.127",
    "nvidia-cuda-runtime-cu12==12.4.127",
    "nvidia-cudnn-cu12==9.1.0.70",
    "nvidia-cufft-cu12==11.2.1.3",
    "nvidia-curand-cu12==10.3.5.147",
    "nvidia-cusolver-cu12==11.6.1.9",
    "nvidia-cusparse-cu12==12.3.1.170",
    "nvidia-cusparselt-cu12==0.6.2",
    "nvidia-nccl-cu12==2.21.5",
    "nvidia-nvjitlink-cu12==12.4.127",
    "nvidia-nvtx-cu12==12.4.127",
    "omegaconf==2.3.0",
    "openunmix==1.3.0",
    "packaging==23.2",
    "pandas==2.2.3",
    "panns-inference==0.1.1",
    "pb-bss-eval==0.0.2",
    "pesq==0.0.4",
    "pillow==11.1.0",
    "platformdirs==4.3.6",
    "pluggy==1.5.0",
    "pooch==1.8.2",
    "prompt_toolkit==3.0.50",
    "propcache==0.3.0",
    "pycparser==2.22",
    "pyparsing==3.2.1",
    "pystoi==0.4.1",
    "pytest==8.3.4",
    "python-dateutil==2.9.0.post0",
    "pytorch-lightning==2.5.0.post0",
    "pytorch-ranger==0.1.1",
    "pytz==2025.1",
    "PyYAML==6.0.2",
    "redis==5.2.1",
    "regex==2024.11.6",
    "requests==2.32.3",
    "retrying==1.3.4",
    "safetensors==0.5.3",
    "scikit-learn==1.6.1",
    "scipy==1.15.2",
    "setuptools==75.8.1",
    "six==1.17.0",
    "soundfile==0.13.1",
    "soxr==0.5.0.post1",
    "sqlparse==0.5.3",
    "submitit==1.5.2",
    "sympy==1.13.1",
    "threadpoolctl==3.5.0",
    "tokenizers==0.21.0",
    "torch==2.6.0",
    "torch-optimizer==0.1.0",
    "torch-stoi==0.2.3",
    "torchaudio==2.6.0",
    "torchlibrosa==0.1.0",
    "torchmetrics==0.11.4",
    "tqdm==4.67.1",
    "transformers==4.49.0",
    "treetable==0.2.5",
    "triton==3.2.0",
    "typing_extensions==4.12.2",
    "tzdata==2025.1",
    "urllib3==2.3.0",
    "vine==5.1.0",
    "wcwidth==0.2.13",
    "yarl==1.18.3",
]

[tool.setuptools]
packages = { find = { where = ["src"] } }